add getUser method

+++ packages/server/servises/User/mockDb.ts
@@ -1,29 +1,61 @@
 import fs from 'fs/promises';
-import { TUser, TUserModel } from './types';
+import { TUser, TUserModel, TUserSearchParams } from './types';
 import { v4 as uuidv4 } from 'uuid';
+
+const message = 'conflict';
+
+type TFindBy = {
+  key: keyof TUserSearchParams;
+  value: string | number;
+};
+
+const checkIsExistingFile = async (path: string) => {
+  try {
+    if (!path) throw Error('path is required');
+
+    await fs.access(path);
+
+    return true;
+  } catch (e) {
+    return false;
+  }
+};
 
 export const DB = {
   async create(user: TUser): Promise<TUserModel | Error> {
-    try {
-      const file = await fs.readFile('./user.json');
-      const parsedFile = JSON.parse(file.toString()) as TUserModel[];
+    const id = uuidv4();
+    const isExist = await checkIsExistingFile('./user.json');
 
-      const findedUser = parsedFile.find((item) => item.email === user.email);
-      if (findedUser) return new Error('user already exist');
-
-      const id = uuidv4();
-
-      fs.writeFile('./user.json', JSON.stringify([...parsedFile, { ...user, id }]));
-
-      return { ...user, id };
-    } catch (e) {
-      const id = uuidv4();
+    if (!isExist) {
       fs.writeFile('./user.json', JSON.stringify([{ ...user, id }]));
-
       return { ...user, id };
     }
 
-    // console.log(file.toString());
-    // fs.writeFile('./user.json', `${Math.random()}`);
+    const file = await fs.readFile('./user.json');
+    const parsedFile = JSON.parse(file.toString()) as TUserModel[];
+
+    const findedUser = parsedFile.find((item) => item.email === user.email);
+    if (findedUser) throw Error(message);
+
+    fs.writeFile('./user.json', JSON.stringify([...parsedFile, { ...user, id }]));
+
+    return { ...user, id };
+  },
+  async findBy(params: TFindBy): Promise<TUserModel[] | Error> {
+    const isExist = await checkIsExistingFile('./user.json');
+
+    if (!isExist) {
+      throw Error('not found');
+    }
+    const file = await fs.readFile('./user.json');
+    const parsedFile = JSON.parse(file.toString()) as TUserModel[];
+
+    const findedUser = parsedFile.filter((item) => item[params.key] === params.value) || [];
+
+    if (findedUser.length) {
+      return findedUser;
+    }
+
+    throw Error('not found');
   },
 };
